+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/java/Eclipse.java
@@ -15,30 +15,33 @@
  */
 package com.diffplug.spotless.maven.java;
 
-import static com.diffplug.spotless.extra.java.EclipseFormatterStep.defaultVersion;
-import static java.util.Collections.singleton;
-
 import java.io.File;
+import java.util.Arrays;
 
 import org.apache.maven.plugins.annotations.Parameter;
 
 import com.diffplug.spotless.FormatterStep;
-import com.diffplug.spotless.extra.java.EclipseFormatterStep;
+import com.diffplug.spotless.extra.EclipseBasedStepBuilder;
+import com.diffplug.spotless.extra.java.EclipseJdtFormatterStep;
 import com.diffplug.spotless.maven.FormatterStepConfig;
 import com.diffplug.spotless.maven.FormatterStepFactory;
 
 public class Eclipse implements FormatterStepFactory {
 
-	@Parameter(required = true)
+	@Parameter
 	private String file;
 
 	@Parameter
 	private String version;
 
 	@Override
-	public FormatterStep newFormatterStep(FormatterStepConfig config) {
-		String formatterVersion = version == null ? defaultVersion() : version;
-		File settingsFile = config.getFileLocator().locateFile(file);
-		return EclipseFormatterStep.create(formatterVersion, singleton(settingsFile), config.getProvisioner());
+	public FormatterStep newFormatterStep(FormatterStepConfig stepConfig) {
+		EclipseBasedStepBuilder eclipseConfig = EclipseJdtFormatterStep.createBuilder(stepConfig.getProvisioner());
+		eclipseConfig.setVersion(version == null ? EclipseJdtFormatterStep.defaultVersion() : version);
+		if (null != file) {
+			File settingsFile = stepConfig.getFileLocator().locateFile(file);
+			eclipseConfig.setPreferences(Arrays.asList(settingsFile));
+		}
+		return eclipseConfig.build();
 	}
 }